check salvataggio prenotazione senza varianti

+++ code/app/Http/Controllers/BookingHandler.php
@@ -112,16 +112,15 @@ class BookingHandler extends Controller
                                     $bpc->value_id = $value_id;
                                     $bpc->save();
                                 }
-
-                                $saved_variants[] = $bpv->id;
-                            } else {
+                            }
+                            else {
                                 if ($bpv->$param != $q) {
                                     $bpv->$param = $q;
                                     $bpv->save();
                                 }
-
-                                $saved_variants[] = $bpv->id;
                             }
+
+                            $saved_variants[] = $bpv->id;
 
                             if ($delivering) {
                                 $bpv->final_price = $bpv->deliveredValue();
@@ -129,6 +128,12 @@ class BookingHandler extends Controller
                             }
 
                             $quantity += $q;
+                        }
+
+                        if ($quantity != 0 && empty($saved_variants)) {
+                            Log::error('Prodotto con varianti, prenotazione senza varianti salvate');
+                            Log::debug("Dump della richiesta:\n" . print_r($request->all(), true));
+                            return $this->errorResponse('Errore nel salvataggio');
                         }
 
                         BookedProductVariant::where('product_id', '=', $booked->id)->whereNotIn('id', $saved_variants)->delete();
